server/requirements.txt
Django==5.1.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
pyjwt==2.8.0
cryptography==42.0.2